+++ workshop-registration/app/__init__.py
@@ -1,0 +1,3 @@
+from .app import AppModel, App 
+from .workshoprepo import WorkshopRepo, WorkshopRecord
+from .registrationrepo import RegistrationRecord, RegistrationRepo
