+++ workshop-registration/external/workshoprepo_inmemory.py
@@ -1,0 +1,12 @@
+from app import WorkshopRepo, WorkshopRecord
+
+
+class InMemoryWorkshopRepo(WorkshopRepo):
+    def __init__(self) -> None:
+        self.workshops = []
+
+    def add_workshop(self, workshop: WorkshopRecord):
+        self.workshops.append(workshop)
+
+    def get_upcoming_workshops(self) -> list[WorkshopRecord]:
+        return self.workshops
